refactor(foldkit): move the renderer out of the boot closure
`startWith` kept the current vnode and the hydration root, restored the
container at teardown, rendered the crash view, built the `Dispatch`
service, ran the view and the patch, scheduled render frames with
`requestAnimationFrame`, wrapped them in View Transitions, and repainted
for DevTools time travel, all inline around the update step that drives
them.

`makeRenderer` in `renderer.ts` now owns all of that. It takes the
render config (`view`, the html builder, `manageDocument`, `crash`, the
slow view and patch configs, the duplicate id scanner, and the resolved
`viewTransition`), the boot's status object, container, build id, init
Model, hydration root, commit notifier, and runtime context, a
`readLiveModel` accessor for the Model the update step owns, and the
Message queue and DevTools integration records whole. It creates the
boundary registry itself, since only the render side uses it. It
returns `crashWith`, `render` for the first paint,
`setLastDirtyMessage`, `scheduleRenderFrame`, `skipPendingViewTransition`,
and the `devToolsRenderBridge` the boot hands to `installDevToolsStore`.

The unit is built where the teardown finalizer used to be registered,
right after the navigation listeners attach and before any fiber forks,
so the finalizer keeps its place in the scope and still runs after the
fibers are interrupted. The runtime context that used to be captured
after the init render is captured just before the unit is built; nothing
in between changed it. The queue and DevTools bindings are now the whole
records, and the boot destructures only what it still uses. The queue
binding gained a type annotation because its `crashWith` wrapper calls
the renderer's and the renderer takes the queue, so TypeScript could not
infer either without it.

Inside the unit the moved code is unchanged apart from formatting, the
four `liveModel` reads that became `readLiveModel()` calls, the two slow
configs under their `maybe` names, the `PendingViewTransition` type
hoisted to module level, the `ResolvedViewTransition` type that moved
with it and is imported back, the context `render` captures renamed to
`renderContext` so it no longer shadows the unit's parameter, and the
`installDevToolsStore` call becoming the `devToolsRenderBridge` value.
The dirty Message write in the update step became `setLastDirtyMessage`.
The `DuplicateIdScanner` type is exported so the unit can name it.

No user-facing effect, so the changeset is empty.
